basic/requirements.txt
google-genai
python-dotenv
groq
ollama
kokoro-tts
soundfile
pyaudio
pygame
SpeechRecognition
